arch-arm, dev-arm: Remove Python 2 compatibility code
Remove uses of six and imports from __future__ and use native Python 3
functionality instead.

Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5/+/39580
Maintainer: Giacomo Travaglini <giacomo.travaglini@arm.com>
Tested-by: kokoro <noreply+kokoro@google.com>

+++ src/arch/arm/ArmSystem.py
@@ -108,7 +108,7 @@ class ArmSystem(System):
         # root instead of appended.
 
         def generateMemNode(mem_range):
-            node = FdtNode("memory@%x" % long(mem_range.start))
+            node = FdtNode("memory@%x" % int(mem_range.start))
             node.append(FdtPropertyStrings("device_type", ["memory"]))
             node.append(FdtPropertyWords("reg",
                 state.addrCells(mem_range.start) +
